Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
langchain
langchain-core
langchain-openai
pydantic
pandas
numpy
scikit-learn
tqdm
